requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.5
django-admin-interface==0.14.2
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-livereload-server==0.3.2
django-phone-field==1.8.1
pytz==2020.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tornado==6.1
